+++ extjvm/action_java_method_delay.go
@@ -8,6 +8,7 @@ import (
 	"context"
 	"github.com/steadybit/action-kit/go/action_kit_api/v2"
 	"github.com/steadybit/action-kit/go/action_kit_sdk"
+	"github.com/steadybit/extension-jvm/extjvm/attack"
 	"github.com/steadybit/extension-kit/extbuild"
 	"github.com/steadybit/extension-kit/extutil"
 	"time"
@@ -20,6 +21,7 @@ type JavaMethodDelayState struct {
 	DelayJitter bool
 	ClassName   string
 	MethodName  string
+	Validate    bool
 	*AttackState
 }
 
@@ -111,6 +113,15 @@ func (l *javaMethodDelay) Describe() action_kit_api.ActionDescription {
 				Required:     extutil.Ptr(true),
 				Advanced:     extutil.Ptr(true),
 			},
+			{
+				Name:         "validate",
+				Label:        "Validate class and method name",
+				Description:  extutil.Ptr("Should the action fail if the specified class and method could not be found?"),
+				Type:         action_kit_api.Boolean,
+				DefaultValue: extutil.Ptr("true"),
+				Required:     extutil.Ptr(true),
+				Advanced:     extutil.Ptr(true),
+			},
 		},
 		Stop: extutil.Ptr(action_kit_api.MutatingEndpointReference{}),
 	}
@@ -140,6 +151,7 @@ func (l *javaMethodDelay) Prepare(_ context.Context, state *JavaMethodDelayState
 	state.DelayJitter = delayJitter
 	state.ClassName = extutil.ToString(request.Config["className"])
 	state.MethodName = extutil.ToString(request.Config["methodName"])
+	state.Validate = extutil.ToBool(request.Config["validate"])
 
 	errResult = extractPid(request, state.AttackState)
 	if errResult != nil {
@@ -160,7 +172,22 @@ func (l *javaMethodDelay) Prepare(_ context.Context, state *JavaMethodDelayState
 // You can mutate the state here.
 // You can use the result to return messages/errors/metrics or artifacts
 func (l *javaMethodDelay) Start(_ context.Context, state *JavaMethodDelayState) (*action_kit_api.StartResult, error) {
-	return commonStart(state.AttackState)
+	result, err := commonStart(state.AttackState)
+	if err != nil {
+		return result, err
+	}
+	if state.Validate {
+		status := attack.GetAttackStatus(state.Pid)
+		if status.AdviceApplied != "APPLIED" {
+			return &action_kit_api.StartResult{
+				Error: extutil.Ptr(action_kit_api.ActionKitError{
+					Title:  "The given class and method did not match anything in the target JVM.",
+					Status: extutil.Ptr(action_kit_api.Failed),
+				}),
+			}, nil
+		}
+	}
+	return result, err
 }
 
 // Stop is called to stop the action