+++ javaagents/attack-java-javaagent/src/main/java/com/steadybit/attacks/javaagent/instrumentation/JavaMethodDelayInstrumentation.java
@@ -27,8 +27,8 @@ public class JavaMethodDelayInstrumentation extends AbstractJavaMethodInstrument
     }
 
     @Override
-    protected AgentBuilder doInstall(AgentBuilder agentBuilder, ElementMatcher.Junction<? super TypeDescription> typeMatcher,
-                                     ElementMatcher.Junction<? super MethodDescription> methodMatcher) {
+    protected AgentBuilder doInstall(AgentBuilder agentBuilder, ElementMatcher<? super TypeDescription> typeMatcher,
+                                     ElementMatcher<? super MethodDescription> methodMatcher) {
         return agentBuilder.type(typeMatcher) //
                 .transform(new AgentBuilder.Transformer.ForAdvice(Advice.withCustomMapping() //
                         .bind(Delay.class, this.delay) //